+++ release-raccoon-app/src/main/java/com/raccoon/dto/ArtistDto.java
@@ -29,4 +29,6 @@ public class ArtistDto {
 
     private String musicbrainzId;
 
+    private Integer followerCount;
+
 }