+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## ⚙️ All important changes to this project are tracked here:
 
-## [Unreleased](https://github.com/jaivalis/release-raccoon/compare/0.4.5...jdevelop)
+## [Unreleased](https://github.com/jaivalis/release-raccoon/compare/0.4.7...jdevelop)
 
+- Feat: Improve search result rankings
+- Feat: Artist follower count
+- Feat: /me/followed-artists pagination
+- Fix: API improvements for integration with UI project
 
 ## [0.4.7](https://github.com/jaivalis/release-raccoon/compare/0.4.6...0.4.7)